requirements.txt
flask==2.3.3
flask-socketio==5.3.4
eventlet==0.33.3
